New DateTime channel, clear response, more retries, update README
Added channel timestampLastUpdate which specifies the time on which the
last successful fetching of data from burner occured.

Did a quick fix of a problem with the responses being aggregated over
time, instead of cleared for each execution of getAllData. This caused
the same data to always be reported to the channels.
A proper solution needs to be done.

Updated README.md.

Instead of only retrying once when a call times out, we now try up to
four times. This should help with not getting the Thing set to OFFLINE
as one retry is not enough in some cases.

Signed-off-by: Jonas Overgaard <freezerz@ubergard.net>